+++ Movieseek/MovieseekAPI/Controllers/UsersController.cs
@@ -34,9 +34,45 @@ namespace MovieseekAPI.Controllers
             _appSettings = appSettings.Value;
         }
 
+        /// <summary>
+        /// Gets all users
+        /// </summary>
+        /// <response code="201">Returns the users</response>
+        /// <response code="401">Unauthorized</response>            
+        [HttpGet]
+        public IActionResult GetAll()
+        {
+            var users = _userService.GetAll();
+            var model = _mapper.Map<IList<UserModel>>(users);
+            return Ok(model);
+        }
+
+        /// <summary>
+        /// Get a user by id
+        /// </summary>
+        /// <response code="200">Returns the user</response>
+        /// <response code="401">Unauthorized</response>            
+        /// <response code="404">User wasn't found</response>            
+        [HttpGet("{id}")]
+        public IActionResult GetById(int id)
+        {
+            var user = _userService.GetById(id);
+            if (user != null)
+            {
+                var model = _mapper.Map<UserModel>(user);
+                return Ok(model);
+            }
+            return NotFound();
+        }
+
+        /// <summary>
+        /// Authenticate a user
+        /// </summary>
+        /// <response code="200">Returns the user with access token</response>
+        /// <response code="400">Bad request</response>            
         [AllowAnonymous]
         [HttpPost("authenticate")]
-        public IActionResult Authenticate([FromBody] AuthenticateModel model)
+        public IActionResult Authenticate([FromBody] AuthenticateUserModel model)
         {
             var user = _userService.Authenticate(model.Username, model.Password);
 
@@ -68,9 +104,14 @@ namespace MovieseekAPI.Controllers
             });
         }
 
+        /// <summary>
+        /// Creates a user
+        /// </summary>
+        /// <response code="201">Returns the created user</response>
+        /// <response code="400">Bad request</response>            
         [AllowAnonymous]
         [HttpPost("register")]
-        public IActionResult Register([FromBody] RegisterModel model)
+        public IActionResult Register([FromBody] RegisterUserModel model)
         {
             if (!ModelState.IsValid)
             {
@@ -94,24 +135,15 @@ namespace MovieseekAPI.Controllers
             }
         }
 
-        [HttpGet]
-        public IActionResult GetAll()
-        {
-            var users = _userService.GetAll();
-            var model = _mapper.Map<IList<UserModel>>(users);
-            return Ok(model);
-        }
-
-        [HttpGet("{id}")]
-        public IActionResult GetById(int id)
-        {
-            var user = _userService.GetById(id);
-            var model = _mapper.Map<UserModel>(user);
-            return Ok(model);
-        }
-
+        /// <summary>
+        /// Updates an user
+        /// </summary>
+        /// <response code="200">Returns the updated user</response>
+        /// <response code="400">Bad request</response>            
+        /// <response code="401">Unauthorized</response>            
+        /// <response code="404">Movie wasn't found</response>            
         [HttpPut("{id}")]
-        public IActionResult Update(int id, [FromBody] UpdateModel model)
+        public IActionResult Update(int id, [FromBody] UpdateUserModel model)
         {
             if (!ModelState.IsValid)
             {
@@ -127,7 +159,8 @@ namespace MovieseekAPI.Controllers
             {
                 // update user 
                 _userService.Update(user, model.Password);
-                return Ok();
+                var updatedUser = _userService.GetById(id);
+                return Ok(updatedUser);
             }
             catch (AppException ex)
             {
@@ -136,11 +169,22 @@ namespace MovieseekAPI.Controllers
             }
         }
 
+        /// <summary>
+        /// Deletes an user
+        /// </summary>
+        /// <response code="204"></response>
+        /// <response code="401">Unauthorized</response>            
+        /// <response code="404">User wasn't found</response>            
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            _userService.Delete(id);
-            return Ok();
+            var user = _userService.GetById(id);
+            if(user != null)
+            {
+                _userService.Delete(id);
+                return NoContent();
+            }
+            return NotFound();
         }
     }
 }
